asdad

+++ TanPhucShopApi/Models/InvoiceDetail.cs
@@ -3,10 +3,9 @@
     public class InvoiceDetail
     {
         public int Id { get; set; }
-        public double Amount { get; set; }
+        public int Amount { get; set; }
         public int ProductId { get; set; }
         public int InvoiceId { get; set; }
-
         public virtual Product Product { get; set; }
         public virtual Invoice Invoice { get; set; }
     }
